requirements.txt
awesome-slugify==1.6.2
Babel==1.3
Flask==0.10.1
git+https://github.com/USStateDept/flask-admin.git@DOSv.11
Flask-BabelEx==0.9.2
Flask-Cache==0.13.1
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-Testing==0.4.2
Flask-Uploads==0.1.3
Flask-Mail==0.9.1
Flask-WTF==0.11
FormEncode==1.2.6
FormEncode-Jinja2==0.1.2
Jinja2==2.7.3
SQLAlchemy==0.9.9
Unidecode==0.04.9
WebHelpers==1.3
anyjson==0.3.3
celery==3.1.17
colander==0.9.3
decorator==3.4.0
economics==0.1.2
lxml==3.4.1
messytables==0.2.1
mistune==0.7
mock==1.0.1
nose==1.3.4
ordereddict==1.1
osvalidate==2014-12-21.01
requests==2.5.1
sqlalchemy-migrate==0.9.2
python-slugify==0.1.0
Flask-Assets==0.10
cssmin==0.2.0
git+https://github.com/USStateDept/cubes.git@DOSv.15
expressions==0.2.2
grako==3.5.1
restpager==0.1
git+https://github.com/kroman0/refine-client-py.git@master
git+https://github.com/miracle2k/webassets@efd904515aa21f165a6d36929fe1ca5684ccba31
git+https://github.com/syrusakbary/pyjade.git@4dbd2ffd7ccdf20c7de6ad7dd8e6b0942bee50eb
https://github.com/USStateDept/Flask-WhooshAlchemy/archive/v.60.zip
xlrd==0.9.3
openpyxl==2.2.2
jsmin==2.1.1
Pygments==2.0.2
Flask-DebugToolbar==0.10.0
